--- GenericsHomework/.editorconfig
@@ -1,4 +1,0 @@
-﻿[*.cs]
-
-# IDE0058: Expression value is never used
-csharp_style_unused_value_expression_statement_preference = unused_local_variable
