--- GenericsHomework/GenericsHomework.Tests/GenericsHomework.Tests.csproj
@@ -1,31 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
-		<IsPackable>false</IsPackable>
-		<IsTestProject>true</IsTestProject>
-	</PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
-	  <PackageReference Include="Moq" Version="4.20.70" />
-	  <PackageReference Include="MSTest.TestAdapter" Version="2.1.2" />
-	  <PackageReference Include="MSTest.TestFramework" Version="2.1.2" />
-	  <PackageReference Include="coverlet.collector" Version="6.0.0">
-		  <PrivateAssets>all</PrivateAssets>
-		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-	  </PackageReference>
-	  <PackageReference Include="xunit" Version="2.6.6" />
-	  <PackageReference Include="xunit.runner.visualstudio" Version="2.5.6">
-		  <PrivateAssets>all</PrivateAssets>
-		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-	  </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\GenericsHomework\GenericsHomework.csproj" />
-  </ItemGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-		<NoWarn>1701;1702;CA1707</NoWarn>
-	</PropertyGroup>
-
-</Project>
