--- GenericsHomework/GenericsHomework.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.8.34408.163
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GenericsHomework", "GenericsHomework\GenericsHomework.csproj", "{7862A6FD-760A-4BB7-A271-9655CBDAA444}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GenericsHomework.Tests", "GenericsHomework.Tests\GenericsHomework.Tests.csproj", "{D545CDA0-AF5A-4CDF-828B-075DBCEE05EA}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D02CEB13-B3B3-4864-97CC-1EFDB7216560}"
-	ProjectSection(SolutionItems) = preProject
-		Directory.Build.props = Directory.Build.props
-		Directory.Build.targets = Directory.Build.targets
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7862A6FD-760A-4BB7-A271-9655CBDAA444}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7862A6FD-760A-4BB7-A271-9655CBDAA444}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7862A6FD-760A-4BB7-A271-9655CBDAA444}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7862A6FD-760A-4BB7-A271-9655CBDAA444}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D545CDA0-AF5A-4CDF-828B-075DBCEE05EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D545CDA0-AF5A-4CDF-828B-075DBCEE05EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D545CDA0-AF5A-4CDF-828B-075DBCEE05EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D545CDA0-AF5A-4CDF-828B-075DBCEE05EA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8CA57E2A-D2D0-4493-8783-695BCBAA5E07}
-	EndGlobalSection
-EndGlobal
