--- global.json
@@ -1,7 +1,0 @@
-{
-  "tools": {
-    "vs": {
-      "version": "16.8"
-    }
-  }
-}
